// route-lists-ui/src/app/models/RouteListOperation.ts
export interface RouteListOperation {
    id: number
    code: string
    name: string
    department: number
    description: string
    labor: number
    position: number
    count: number
    type: number
    startDate: string
    endDate: string
    number: string
    routeListId: number
    executorId: number,
    executor: string
}